最新版本

+++ src/main/java/yidong/service/RetailService.java
@@ -6,9 +6,9 @@ import java.util.List;
 import java.util.Map;
 
 public interface RetailService {
-    int getDay(String time);
-    int getWeek(String time);
-    int getMonth(String time);
+    Float getDay(String time);
+    Float getWeek(String time);
+    Float getMonth(String time);
     List<Retail> selectRetail(Map map);
     int selectCount(Map map);
 }
